rename a few cron file
